# backEnd/swagger.yaml
openapi: 3.0.3
info:
  title: Subastas G16 API
  version: "1.0"
  description: |
    Backend del sistema de subastas — UADE DA1 Grupo 16 (rama **dev**)

    ## Autenticación
    La mayoría de endpoints requieren un **Bearer JWT** en el header `Authorization`.
    Primero hacé login con `POST /api/auth/login` y usá el `accessToken` devuelto.
servers:
  - url: http://localhost:8080
    description: Servidor local (dev)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AprobacionRequest:
      type: object
      required: [categoria]
      properties:
        categoria:
          type: string
          enum: [comun, especial, plata, oro, platino]
          example: plata
...
    MedioPagoRequest:
      type: object
      required: [tipo, moneda]
      properties:
        tipo:
          type: string
          enum: [CUENTA_BANCARIA, TARJETA_CREDITO, CHEQUE_CERTIFICADO]
          example: CUENTA_BANCARIA
        moneda:
          type: string
          enum: [ARS, USD]
          example: ARS
        # Datos de Cuenta Bancaria
        banco:
          type: string
          example: Banco Nación
        numeroCuenta:
          type: string
          example: "0012345678"
        cbuSwift:
          type: string
          example: "0340099801234567890123"
        esInternacional:
          type: boolean
          example: false
        # Datos de Tarjeta de Crédito
        numeroTarjeta:
          type: string
          example: "4111111111111111"
        titular:
          type: string
          example: JUAN PEREZ
        vencimiento:
          type: string
          example: "12/27"
        esTarjetaInternacional:
          type: boolean
          example: false
        # Datos de Cheque Certificado
        numeroCheque:
          type: string
          example: "00123456"
        bancoEmisor:
          type: string
          example: Banco Galicia
        montoCertificado:
          type: number
          format: decimal
          example: 50000.00
...
    ClienteResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Juan Pérez
        documento:
          type: string
          example: "35123456"
        direccion:
          type: string
          example: Av. Corrientes 1234
        pais:
          type: string
          example: Argentina
        email:
          type: string
          example: juan@email.com
        categoria:
          type: string
          example: plata
        estado:
          type: string
          example: APROBADO
        tieneMedioPagoVerificado:
          type: boolean
          example: true
...
    MensajeResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Operación realizada exitosamente
...
paths:
...
  /api/admin/usuarios/{uuid}/aprobar:
    put:
      tags: [Administración]
      summary: Aprobar usuario
      description: |
        Aprueba un usuario por su UUID asignándole una categoría.
        El sistema envía un email de confirmación automáticamente.
      operationId: aprobarUsuario
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AprobacionRequest'
      responses:
        '200':
          description: Usuario aprobado
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Usuario aprobado exitosamente
                  uuid:
                    type: string
                  categoria:
                    type: string
        '404':
          description: Usuario no encontrado
  /api/admin/usuarios/{uuid}/rechazar:
    put:
      tags: [Administración]
      summary: Rechazar usuario
      description: Rechaza un usuario por su UUID. El usuario no podrá hacer login.
      operationId: rechazarUsuario
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        '200':
          description: Usuario rechazado
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Usuario rechazado
        '404':
          description: Usuario no encontrado

  # ──────────── CLIENTES ────────────
  /api/clientes/perfil:
    get:
      tags: [Clientes]
      summary: Obtener perfil del cliente autenticado
      operationId: getPerfilCliente
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '401':
          description: No autenticado
  /api/clientes/metricas:
    get:
      tags: [Clientes]
      summary: Obtener métricas de participación
      description: Devuelve el total de pujas realizadas, victorias y tasa de éxito del cliente autenticado.
      operationId: getMetricasCliente
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Métricas del cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPujas:
                    type: integer
                    example: 35
                  victorias:
                    type: integer
                    example: 8
                  tasaExito:
                    type: number
                    format: double
                    example: 22.86
        '401':
          description: No autenticado
  /api/clientes/info-subasta:
    get:
      tags: [Clientes]
      summary: Info previa a ingresar a una subasta
      description: |
        Verifica si el cliente puede participar en subastas:
        - Si tiene multas pendientes
        - Si tiene al menos un medio de pago verificado
      operationId: getInfoParaSubasta
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Info de habilitación
          content:
            application/json:
              schema:
                type: object
                properties:
                  puedeParticipar:
                    type: boolean
                    example: true
                  tieneMedioPago:
                    type: boolean
                    example: true
                  tieneMultasPendientes:
                    type: boolean
                    example: false
                  multasPendientes:
                    type: integer
                    example: 0
        '401':
          description: No autenticado

  # ──────────── SUBASTAS ────────────
...
  /api/medios-pago:
    post:
      tags: [Medios de Pago]
      summary: Registrar medio de pago
      description: Registra un nuevo medio de pago para el cliente autenticado.
      operationId: registrarMedioPago
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedioPagoRequest'
      responses:
        '201':
          description: Medio de pago creado (pendiente de verificación)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 7
                  tipo:
                    type: string
                    example: CUENTA_BANCARIA
                  moneda:
                    type: string
                    example: ARS
                  verificado:
                    type: boolean
                    example: false
        '401':
          description: No autenticado

    get:
      tags: [Medios de Pago]
      summary: Listar mis medios de pago
      description: Devuelve todos los medios de pago activos del cliente autenticado.
      operationId: listarMediosPago
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de medios de pago
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    tipo:
                      type: string
                    moneda:
                      type: string
                    verificado:
                      type: boolean
                    activo:
                      type: boolean
  /api/medios-pago/{id}:
    delete:
      tags: [Medios de Pago]
      summary: Eliminar medio de pago
      description: Desactiva (soft delete) un medio de pago del cliente autenticado.
      operationId: eliminarMedioPago
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 7
      responses:
        '200':
          description: Medio de pago eliminado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeResponse'
        '404':
          description: Medio de pago no encontrado
  /api/medios-pago/admin/{id}/verificar:
    put:
      tags: [Medios de Pago]
      summary: "[ADMIN] Verificar medio de pago"
      description: Marca un medio de pago como verificado. Permite al cliente participar en subastas.
      operationId: verificarMedioPago
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 7
      responses:
        '200':
          description: Medio de pago verificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeResponse'

  # ──────────── MULTAS ────────────
  /api/multas:
    get:
      tags: [Multas]
      summary: Listar todas las multas (pagadas y pendientes)
      operationId: listarTodasMultas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista completa de multas del cliente autenticado
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    subastaId:
                      type: integer
                    itemId:
                      type: integer
                    montoOfertado:
                      type: number
                    montoMulta:
                      type: number
                    pagada:
                      type: boolean
                    fechaLimite:
                      type: string
                      format: date
  /api/multas/pendientes:
    get:
      tags: [Multas]
      summary: Listar multas pendientes de pago
      operationId: listarMultasPendientes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Multas pendientes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    montoMulta:
                      type: number
                    fechaLimite:
                      type: string
                      format: date
  /api/multas/{id}/pagar:
    post:
      tags: [Multas]
      summary: Pagar una multa
      description: Registra el pago de la multa. Habilita al cliente a participar nuevamente en subastas.
      operationId: pagarMulta
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 3
      responses:
        '200':
          description: Multa pagada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Multa pagada correctamente
                  multaId:
                    type: integer
        '404':
          description: Multa no encontrada
  /api/multas/admin/generar:
    post:
      tags: [Multas]
      summary: "[ADMIN] Generar multa manual"
      description: Genera una multa para un cliente que no pagó su puja ganada.
      operationId: generarMultaManual
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clienteId, subastaId, itemId, montoOfertado]
              properties:
                clienteId:
                  type: integer
                  example: 1
                subastaId:
                  type: integer
                  example: 10
                itemId:
                  type: integer
                  example: 42
                montoOfertado:
                  type: number
                  example: 15000.00
      responses:
        '200':
          description: Multa generada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Multa generada exitosamente
                  multaId:
                    type: integer
                    format: int64
                  montoMulta:
                    type: number
                    example: 3000.00
                  fechaLimite:
                    type: string
                    format: date
                    example: "2025-06-22"

  # ──────────── NOTIFICACIONES ────────────
...